+++ src/main/java/com/analytique/gestion_analytique/Models/ContratEmployeId.java
@@ -3,19 +3,19 @@ package com.analytique.gestion_analytique.Models;
 import java.io.Serializable;
 
 public class ContratEmployeId implements Serializable {
-    private Long contrat;
-    private Long employe;
+    private Integer contrat;
+    private Integer employe;
 
-    public Long getContrat() {
+    public Integer getContrat() {
         return contrat;
     }
-    public void setContrat(Long contrat) {
+    public void setContrat(Integer contrat) {
         this.contrat = contrat;
     }
-    public Long getEmploye() {
+    public Integer getEmploye() {
         return employe;
     }
-    public void setEmploye(Long employe) {
+    public void setEmploye(Integer employe) {
         this.employe = employe;
     }
